Add files via upload
Introduction:
With the growing popularity of streaming services like Netflix, it is more important than ever for companies to understand their audience and optimize their content offerings. This project aims to leverage data analysis to understand the type, variety and magnitude of Movies and TV Shows released on Netflix till 2020. 
Methodology: 
To achieve this goal, I have taken data from Kaggle and dashboard is prepared for highlighting different points in Netflix style view which you can see in this project.